moved main logger
