+++ spatialprofilingtoolbox/__init__.py
@@ -21,7 +21,7 @@ from .workflow.workflows import get_workflow
 from .workflow.workflows import get_workflow_names
 from .workflow.environment.configuration_settings import get_version
 
-from .log_formats import colorized_logger
+from .standalone_utilities.log_formats import colorized_logger
 logger = colorized_logger(__name__)
 
 __version__ = get_version()
