+++ spatialprofilingtoolbox/countsserver/scripts/cache-expressions-data-array.py
@@ -11,7 +11,7 @@ import json
 def do_library_imports():
     import spatialprofilingtoolbox
     from spatialprofilingtoolbox.db.database_connection import DatabaseConnectionMaker
-    from spatialprofilingtoolbox.log_formats import colorized_logger
+    from spatialprofilingtoolbox.standalone_utilities.log_formats import colorized_logger
     logger = colorized_logger('cache-expressions-data-array')
 
     from spatialprofilingtoolbox.standalone_utilities.module_load_error import SuggestExtrasException
