+++ spatialprofilingtoolbox/countsserver/scripts/start.py
@@ -4,7 +4,7 @@ import socketserver
 import os
 import re
 
-from log_formats import colorized_logger
+from.standalone_utilities.log_formats import colorized_logger
 logger = colorized_logger('start')
