+++ spatialprofilingtoolbox/workflow/dataset_designs/multiplexed_imaging/halo_cell_metadata_design.py
@@ -6,7 +6,7 @@ import re
 
 import pandas as pd
 
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 
 logger = colorized_logger(__name__)
