+++ spatialprofilingtoolbox/workflow/environment/database_context_utility.py
@@ -1,7 +1,7 @@
 import sqlite3
 import time
 
-from ...log_formats import colorized_logger
+from ...standalone_utilities.log_formats import colorized_logger
 
 logger = colorized_logger(__name__)
