+++ spatialprofilingtoolbox/workflow/environment/export_features.py
@@ -4,7 +4,7 @@ import pandas as pd
 
 from .source_file_parsers.parser import SourceToADIParser
 from ...db.database_connection import DatabaseConnectionMaker
-from ...log_formats import colorized_logger
+from ...standalone_utilities.log_formats import colorized_logger
 logger = colorized_logger(__name__)
