+++ spatialprofilingtoolbox/workflow/environment/file_io.py
@@ -9,7 +9,7 @@ from itertools import repeat
 
 import pandas as pd
 
-from ...log_formats import colorized_logger
+from ...standalone_utilities.log_formats import colorized_logger
 logger = colorized_logger(__name__)
