+++ spatialprofilingtoolbox/workflow/environment/source_file_parsers/channels.py
@@ -4,7 +4,7 @@ from os.path import join
 import pandas as pd
 
 from .parser import SourceToADIParser
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 logger = colorized_logger(__name__)
