+++ spatialprofilingtoolbox/workflow/environment/source_file_parsers/parser.py
@@ -4,7 +4,7 @@ import enum
 from enum import Enum
 from enum import auto
 
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 logger = colorized_logger(__name__)
