+++ spatialprofilingtoolbox/workflow/environment/source_file_parsers/skimmer.py
@@ -5,7 +5,7 @@ import psycopg2
 from psycopg2 import sql
 import pandas as pd
 
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 logger = colorized_logger(__name__)
 
 from ....database_connection import DatabaseConnectionMaker
