+++ spatialprofilingtoolbox/workflow/environment/source_file_parsers/subjects.py
@@ -2,7 +2,7 @@
 import pandas as pd
 
 from .parser import SourceToADIParser
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 logger = colorized_logger(__name__)
