+++ spatialprofilingtoolbox/workflow/environment/verbose_sql_execution.py
@@ -1,6 +1,6 @@
 import importlib.resources
 
-from ...log_formats import colorized_logger
+from ...standalone_utilities.log_formats import colorized_logger
 logger = colorized_logger(__name__)
