+++ spatialprofilingtoolbox/workflow/workflows/defaults/computational_design.py
@@ -1,7 +1,7 @@
 
 import pandas as pd
 
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 
 logger = colorized_logger(__name__)
