+++ spatialprofilingtoolbox/workflow/workflows/defaults/core.py
@@ -7,7 +7,7 @@ import pandas as pd
 from ...environment.file_io import raw_line_count
 from ...environment.dichotomization import Dichotomizer
 from ...environment.logging.performance_timer import PerformanceTimer
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 
 logger = colorized_logger(__name__)
