+++ spatialprofilingtoolbox/workflow/workflows/density/core.py
@@ -5,7 +5,7 @@ import pandas as pd
 
 from ...environment.database_context_utility import WaitingDatabaseContextManager
 from ..defaults.core import CoreJob
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 from .data_logging import DensityDataLogger
 
 logger = colorized_logger(__name__)
