+++ spatialprofilingtoolbox/workflow/workflows/front_proximity/core.py
@@ -8,7 +8,7 @@ from scipy.spatial import KDTree
 
 from ..defaults.core import CoreJob
 from ...environment.database_context_utility import WaitingDatabaseContextManager
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 
 logger = colorized_logger(__name__)
