+++ spatialprofilingtoolbox/workflow/workflows/front_proximity/integrator.py
@@ -1,5 +1,5 @@
 
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 
 logger = colorized_logger(__name__)
