+++ spatialprofilingtoolbox/workflow/workflows/halo_import/core.py
@@ -1,7 +1,7 @@
 import sqlite3
 
 from ..defaults.core import CoreJob
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 
 logger = colorized_logger(__name__)
