+++ spatialprofilingtoolbox/workflow/workflows/halo_import/initializer.py
@@ -2,7 +2,7 @@
 from ..defaults.initializer import Initializer
 from ...environment.configuration_settings import default_db_config_filename
 from ...environment.source_file_parsers.skimmer import DataSkimmer
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 
 logger = colorized_logger(__name__)
