+++ spatialprofilingtoolbox/workflow/workflows/nearest_distance/integrator.py
@@ -1,4 +1,4 @@
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 
 logger = colorized_logger(__name__)
