+++ spatialprofilingtoolbox/workflow/workflows/phenotype_proximity/core.py
@@ -15,7 +15,7 @@ from sklearn.neighbors import BallTree
 
 from ...environment.database_context_utility import WaitingDatabaseContextManager
 from ..defaults.core import CoreJob
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 from .computational_design import PhenotypeProximityDesign
 
 logger = colorized_logger(__name__)
