+++ spatialprofilingtoolbox/workflow/workflows/phenotype_proximity/integrator.py
@@ -11,7 +11,7 @@ import numpy as np
 from scipy.stats import ttest_ind
 from scipy.stats import kruskal
 
-from ....log_formats import colorized_logger
+from ....standalone_utilities.log_formats import colorized_logger
 from ...environment.export_features import ADIFeaturesUploader
 from ...environment.source_file_parsers.parser import get_unique_value
 from .computational_design import PhenotypeProximityDesign
